+++ core/src/xmake/engine.c
@@ -37,7 +37,7 @@
 #   include <unistd.h>
 #   include <mach-o/dyld.h>
 #   include <signal.h>
-#elif defined(TB_CONFIG_OS_LINUX) || defined(TB_CONFIG_OS_BSD) || defined(TB_CONFIG_OS_ANDROID)
+#elif defined(TB_CONFIG_OS_LINUX) || defined(TB_CONFIG_OS_BSD) || defined(TB_CONFIG_OS_ANDROID) || defined(TB_CONFIG_OS_HAIKU)
 #   include <unistd.h>
 #   include <signal.h>
 #endif
@@ -877,6 +877,8 @@ static tb_void_t xm_engine_init_host(xm_engine_t* engine)
     syshost = "ios";
 #elif defined(TB_CONFIG_OS_ANDROID)
     syshost = "android";
+#elif defined(TB_CONFIG_OS_HAIKU)
+    syshost = "haiku";
 #endif
     lua_pushstring(engine->lua, syshost? syshost : "unknown");
     lua_setglobal(engine->lua, "_HOST");