+++ xmake/platforms/bsd/xmake.lua
@@ -40,7 +40,7 @@ platform("bsd")
     set_installdir("/usr/local")
 
     -- set toolchains
-    set_toolchains("envs", "gcc", "clang", "yasm", "nasm", "fasm", "cuda", "dlang", "go", "rust", "gfortran", "zig")
+    set_toolchains("envs", "gcc", "clang", "yasm", "nasm", "fasm", "cuda", "go", "rust", "gfortran", "zig")
 
     -- set menu
     set_menu {