+++ xmake/plugins/format/main.lua
@@ -32,7 +32,7 @@ function main()
     -- load configuration
     config.load()
 
-    -- enter the environments of doxygen
+    -- enter the environments of llvm
     local oldenvs = packagenv.enter("llvm")
 
     -- find clang-format
@@ -73,10 +73,9 @@ function main()
 
     -- set file to format
     if option.get("files") then
-        -- list of files/path to format
-        local files = option.get("file"):split(" ")
+        local files = path.splitenv(option.get("files"))
         for _, f in ipairs(files) do
-            local p = path.join(project, f)
+            local p = path.join(projectdir, f)
             for _, filepath in ipairs(os.files(p)) do
                 table.insert(argv, filepath)
             end
@@ -97,7 +96,8 @@ function main()
 
     -- format files
     os.vrunv(clang_format.program, argv, {curdir = projectdir})
-    cprint("${color.success}formatting complete")
+    cprint("${color.success}format ok!")
 
+    -- done
     os.setenvs(oldenvs)
 end