+++ xmake/rules/c++/build_optimization/xmake.lua
@@ -21,12 +21,12 @@
 -- define rule: c.build.optimization
 rule("c.build.optimization")
     on_config(function (target)
-        import("load")(target, "cc")
+        import("config")(target, "cc")
     end)
 
 -- define rule: c++.build.optimization
 rule("c++.build.optimization")
     on_config(function (target)
-        import("load")(target, "cxx")
+        import("config")(target, "cxx")
     end)