+++ xmake/rules/cuda/env/xmake.lua
@@ -18,12 +18,10 @@
 -- @file        xmake.lua
 --
 
--- define rule: environment
 rule("cuda.env")
 
     on_load(function (target)
         import("detect.sdks.find_cuda")
-
         local cuda = assert(find_cuda(nil, {verbose = true}), "Cuda SDK not found!")
         if cuda then
             target:data_set("cuda", cuda)